--- EModulesApp/app/src/main/java/com/hashinclude/cmoc/emodulesapp/MyAppGlideModule.java
@@ -1,17 +1,0 @@
-package com.hashinclude.cmoc.emodulesapp;
-
-import com.bumptech.glide.annotation.GlideModule;
-import com.bumptech.glide.module.AppGlideModule;
-
-/**
- * Created by harsh on 2/5/18.
- */
-
-/*
-* You’re not required to implement any of the methods in AppGlideModule for the API to be generated.
-* You can leave the class blank as long as it extends AppGlideModule and is annotated with @GlideModule.
-AppGlideModule implementations must always be annotated with @GlideModule.
- If the annotation is not present, the module will not be discovered and you will see a warning in your logs
- with the Glide log tag that indicates that the module couldn’t be found.*/
-@GlideModule
-public class MyAppGlideModule extends AppGlideModule {}
