+++ EModulesApp/app/src/main/java/com/hashinclude/cmoc/emodulesapp/eventListeners/OnItemClickListener.java
@@ -1,4 +1,4 @@
-package com.hashinclude.cmoc.emodulesapp;
+package com.hashinclude.cmoc.emodulesapp.eventListeners;
 
 import android.view.View;
