+++ EModulesApp/app/src/main/java/com/hashinclude/cmoc/emodulesapp/eventListeners/SubmitButtonClickedEvent.java
@@ -1,4 +1,4 @@
-package com.hashinclude.cmoc.emodulesapp;
+package com.hashinclude.cmoc.emodulesapp.eventListeners;
 
 /**
  * Created by harsh on 2/10/18.
