+++ EModulesApp/app/src/main/java/com/hashinclude/cmoc/emodulesapp/utils/Constants.java
@@ -1,4 +1,4 @@
-package com.hashinclude.cmoc.emodulesapp;
+package com.hashinclude.cmoc.emodulesapp.utils;
 
 public class Constants {
     public static int numberOfOptions = 3;
