+++ EModulesApp/app/src/main/res/layout/activity_course_selection.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent" android:layout_height="match_parent">
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:layout_centerVertical="true"
+        android:gravity="center_horizontal"
+        >
+        <Button
+            android:id="@+id/course_1"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="@dimen/margin_large"
+            android:text="CFA"
+            />
+        <Button
+            android:id="@+id/course_2"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="GMAT"
+            />
+    </LinearLayout>
+</RelativeLayout>
