+++ EModulesApp/app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
     sothree:umanoParallaxOffset="100dp"
     sothree:umanoScrollableView="@+id/onScrollView"
     sothree:umanoShadowHeight="4dp"
-    tools:context="com.hashinclude.cmoc.emodulesapp.MainActivity">
+    tools:context="com.hashinclude.cmoc.emodulesapp.activities.MainActivity">
 
     <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
