+++ EModulesApp/app/src/main/res/layout/activity_single_question.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.hashinclude.cmoc.emodulesapp.SingleQuestionActivity">
+    tools:context="com.hashinclude.cmoc.emodulesapp.activities.SingleQuestionActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/questionActivityToolbar"
