+++ EModulesApp/app/src/main/res/layout/fragment_solution.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:background="@android:color/white"
     android:orientation="vertical"
-    tools:context="com.hashinclude.cmoc.emodulesapp.SolutionFragment">
+    tools:context="com.hashinclude.cmoc.emodulesapp.fragments.SolutionFragment">
 
     <LinearLayout
         android:id="@+id/rootWhenSolutionVisible"
